requirements.txt
pyqt5
opencv-python
pillow
google-cloud-storage
google-generativeai 